Extract cold-start live driver into its own module; pin --target-dir
Address PR #1377 codecov/patch failure and Codex P2 review.

- Pass an explicit `--target-dir` (pinned inside the throwaway project) to
  the cold-start `cargo build`. `env_remove` only neutralises env-var
  overrides; a `.cargo/config.toml` `build.target-dir` anywhere up the tree
  could still redirect artifacts into a shared, possibly-warm directory. The
  explicit flag is highest-precedence and guarantees a first clean compile
  into the empty dir.

- Split the cold-start benchmark so coverage reflects what is actually
  testable, matching the repo's existing convention (build.rs / dev.rs):
    * New `cold_start_driver.rs` holds the live measurement orchestration
      (scaffold + cold cargo build + server spawn + HTTP poll) — pure
      subprocess/TCP/filesystem I/O that cannot be unit-tested. Added to the
      codecov `ignore` list with a rationale comment.
    * `dev_loop_bench.rs` keeps the pure, unit-tested logic (budgets, stats,
      budget checking, report building, dry-run table, cargo artifact-path
      parsing), now measured for coverage.

- Add unit tests for the remaining pure helpers: `build_diagnostics` next
  actions for every warm change class, and `cargo_executable_path` skipping
  interleaved non-JSON cargo output.
