+++ autumn-cli/src/dev_loop_bench.rs
@@ -9,9 +9,7 @@
 //! methodology used to measure end-to-end latency.
 
 use std::fmt::Write as FmtWrite;
-use std::path::{Path, PathBuf};
-use std::process::{Command, Stdio};
-use std::time::{Duration, Instant};
+use std::path::PathBuf;
 
 use serde::Serialize;
 
@@ -540,7 +538,7 @@ fn print_budget_table() {
 
 /// Render the cold-start onboarding budget table. The database-backed shape is
 /// included only when `include_db` is set (it is informational in this slice).
-fn format_cold_start_budget_table(include_db: bool) -> String {
+pub fn format_cold_start_budget_table(include_db: bool) -> String {
     let classes: &[ChangeClass] = if include_db {
         &[ChangeClass::ColdStartHello, ChangeClass::ColdStartDb]
     } else {
@@ -592,25 +590,15 @@ fn rust_version_string() -> String {
 }
 
 // ── Cold-start onboarding benchmark (issue #977) ──────────────────────────────
-
-/// Which scaffolded project shape to measure for cold start.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ColdStartShape {
-    /// No-database `hello` shape (the gated onboarding budget).
-    Hello,
-    /// Database-backed shape (informational only in this slice).
-    Db,
-}
-
-const fn class_for(shape: ColdStartShape) -> ChangeClass {
-    match shape {
-        ColdStartShape::Hello => ChangeClass::ColdStartHello,
-        ColdStartShape::Db => ChangeClass::ColdStartDb,
-    }
-}
+//
+// This file holds the **pure, unit-tested** half of the cold-start benchmark:
+// budgets, statistics, budget checking, report building, and the dry-run budget
+// table. The **live measurement driver** (scaffold → cold compile → serve → time
+// the first 200) is all subprocess / TCP / filesystem I/O and lives in
+// [`crate::cold_start_driver`], which is excluded from coverage like `dev.rs`.
 
 /// Outcome of the (optional) database-backed cold-start shape.
-enum DbOutcome {
+pub enum DbOutcome {
     /// `--include-db` was not requested.
     NotRequested,
     /// Measured samples (milliseconds).
@@ -626,7 +614,7 @@ enum DbOutcome {
 /// `hello_samples` always gates the result. The database-backed shape, when
 /// requested, is recorded as **informational** (measured or failed) and never
 /// affects `all_passed`.
-fn build_cold_start_report(hello_samples: &[u64], db: &DbOutcome) -> FullReport {
+pub fn build_cold_start_report(hello_samples: &[u64], db: &DbOutcome) -> FullReport {
     let mut results = Vec::new();
 
     let hello_budget = budget_for(ChangeClass::ColdStartHello);
@@ -687,234 +675,13 @@ fn db_failure_result(msg: &str) -> BudgetCheckResult {
     }
 }
 
-/// Locate the workspace `autumn-web` crate so a throwaway project can be built
-/// against the repository's actual source (a genuine clean compile) rather than
-/// a possibly-unpublished crates.io version.
-///
-/// Honours the `AUTUMN_BENCH_AUTUMN_WEB_PATH` override, otherwise walks up from
-/// the current directory looking for an `autumn/Cargo.toml` whose package is
-/// `autumn-web`.
-fn locate_autumn_web() -> Option<PathBuf> {
-    if let Ok(p) = std::env::var("AUTUMN_BENCH_AUTUMN_WEB_PATH") {
-        let pb = PathBuf::from(p);
-        if pb.join("Cargo.toml").is_file() {
-            return Some(pb);
-        }
-    }
-    let cwd = std::env::current_dir().ok()?;
-    for ancestor in cwd.ancestors() {
-        let candidate = ancestor.join("autumn");
-        let manifest = candidate.join("Cargo.toml");
-        if manifest.is_file()
-            && let Ok(s) = std::fs::read_to_string(&manifest)
-            && s.contains("name = \"autumn-web\"")
-        {
-            return Some(candidate);
-        }
-    }
-    None
-}
-
-/// Append a `[patch.crates-io]` override pointing `autumn-web` at local source.
-///
-/// The scaffolded `Cargo.toml` already declares its own (empty) `[workspace]`
-/// table, so it is a standalone workspace root and a `[patch]` section is valid
-/// there. The patch applies regardless of whether the dependency is the plain
-/// `autumn-web = "x"` form or the daemon/`features` table form.
-fn repoint_autumn_web(project_dir: &Path, autumn_web: &Path) -> Result<(), String> {
-    let manifest = project_dir.join("Cargo.toml");
-    let mut content =
-        std::fs::read_to_string(&manifest).map_err(|e| format!("read Cargo.toml: {e}"))?;
-    // `{:?}` quotes and escapes the path into a valid TOML basic string.
-    let patch = format!(
-        "\n[patch.crates-io]\nautumn-web = {{ path = {:?} }}\n",
-        autumn_web.display().to_string()
-    );
-    content.push_str(&patch);
-    std::fs::write(&manifest, content).map_err(|e| format!("write Cargo.toml: {e}"))?;
-    Ok(())
-}
-
-/// Measure a single cold-start journey end-to-end: scaffold a throwaway project,
-/// compile it from a cold target, start it, and time until the first HTTP 200.
-///
-/// Returns the elapsed milliseconds from just before `autumn new` to the first
-/// successful response on `/`.
-fn measure_cold_start_once(shape: ColdStartShape) -> Result<u64, String> {
-    let autumn_web = locate_autumn_web().ok_or_else(|| {
-        "could not locate the workspace autumn-web crate; \
-         set AUTUMN_BENCH_AUTUMN_WEB_PATH to its directory"
-            .to_string()
-    })?;
-    let autumn_web = std::fs::canonicalize(&autumn_web)
-        .map_err(|e| format!("canonicalize autumn-web path: {e}"))?;
-
-    let tmp = tempfile::tempdir().map_err(|e| format!("create temp dir: {e}"))?;
-    let project_name = "coldstart_app";
-
-    let opts = match shape {
-        // The DB-free daemon starter compiles and serves without Postgres.
-        ColdStartShape::Hello => crate::new::GenerateOptions {
-            with_daemon: true,
-            ..Default::default()
-        },
-        // The database-backed shape uses the bundled managed-Postgres provider so
-        // the app self-provisions a real Postgres (via `postgresql_embedded`),
-        // runs migrations, and connects before serving — a genuine DB-backed
-        // onboarding cold start with no external service required.
-        ColdStartShape::Db => crate::new::GenerateOptions {
-            with_bundled_pg: true,
-            ..Default::default()
-        },
-    };
-
-    // Start the clock just before `autumn new` so the whole onboarding journey
-    // (scaffold → first clean compile → boot → first 200) is captured.
-    let start = Instant::now();
-
-    // Quiet scaffold: keep stdout clean so `--cold-start --json` emits only the
-    // JSON report.
-    crate::new::generate_with_quiet(project_name, tmp.path(), opts)
-        .map_err(|e| format!("autumn new failed: {e}"))?;
-    let project_dir = tmp.path().join(project_name);
-
-    repoint_autumn_web(&project_dir, &autumn_web)?;
-
-    // Pick the port the app binds and we poll. Default: a freshly reserved
-    // ephemeral port per sample, so repeated runs never collide on a lingering
-    // TIME_WAIT socket from the previous sample. An explicit AUTUMN_BENCH_PORT
-    // override is honoured as-is.
-    let port = match std::env::var("AUTUMN_BENCH_PORT")
-        .ok()
-        .and_then(|p| p.parse::<u16>().ok())
-    {
-        Some(p) => p,
-        None => reserve_free_port()?,
-    };
-
-    // Cold build into the project's own (empty) target/. Remove every inherited
-    // Cargo target/triple override so the parent's warm cache is never reused,
-    // and ask Cargo for JSON artifact messages so we learn the *exact* built
-    // binary path — robust to any `build.target` / `build.target-dir` coming from
-    // Cargo config files, which `env_remove` alone cannot neutralise.
-    let output = Command::new("cargo")
-        .args(["build", "--message-format=json"])
-        .current_dir(&project_dir)
-        .env_remove("CARGO_TARGET_DIR")
-        .env_remove("CARGO_BUILD_TARGET_DIR")
-        .env_remove("CARGO_BUILD_TARGET")
-        .output()
-        .map_err(|e| format!("cargo build spawn failed: {e}"))?;
-    if !output.status.success() {
-        return Err(format!(
-            "cargo build failed for the scaffolded project:\n{}",
-            String::from_utf8_lossy(&output.stderr)
-        ));
-    }
-    let bin = cargo_executable_path(&output.stdout, project_name).ok_or_else(|| {
-        "could not determine the built binary path from cargo's JSON output".to_string()
-    })?;
-
-    let mut cmd = Command::new(&bin);
-    cmd.current_dir(&project_dir);
-    // Isolate the child from the surrounding environment: clear every inherited
-    // `AUTUMN_*` var so app-mode settings can't change how the scaffolded app
-    // boots or what it binds. Without this, e.g. `AUTUMN_ENV=prod` would trigger
-    // production boot checks, `AUTUMN_BUILD_STATIC=1` would render static output
-    // and exit instead of serving, and a stray `AUTUMN_SERVER__UNIX_SOCKET`
-    // would make the app bind a socket instead of the TCP port we poll.
-    for (key, _) in std::env::vars() {
-        if key.starts_with("AUTUMN_") {
-            cmd.env_remove(&key);
-        }
-    }
-    // Then pin exactly the host+port we poll (env vars are the highest-precedence
-    // config source, overriding autumn.toml).
-    cmd.env("AUTUMN_SERVER__HOST", "127.0.0.1")
-        .env("AUTUMN_SERVER__PORT", port.to_string());
-    // For the DB shape, keep the bundled managed-Postgres cluster inside the
-    // throwaway temp dir so the run is self-contained: otherwise the provider
-    // defaults to `$HOME/.local/share/autumn/...` and leaves a full Postgres
-    // data/extraction directory behind after the tempdir is removed.
-    if matches!(shape, ColdStartShape::Db) {
-        cmd.env(
-            "AUTUMN_MANAGED_PG_DATA_DIR",
-            project_dir.join("managed-pg-data"),
-        );
-    }
-    cmd
-        // Discard the app's own logs so they never interleave with the JSON
-        // report on stdout; the harness reports progress on stderr separately.
-        .stdout(Stdio::null())
-        .stderr(Stdio::null());
-    let mut child = cmd
-        .spawn()
-        .map_err(|e| format!("failed to start the built server binary: {e}"))?;
-
-    // Poll the configured root route until the first 200 or the deadline.
-    let url = format!("http://127.0.0.1:{port}/");
-    let deadline = start + Duration::from_millis(budget_for(class_for(shape)).max_ms * 2);
-    let client = reqwest::blocking::Client::builder()
-        .timeout(Duration::from_secs(2))
-        .build()
-        .map_err(|e| format!("build http client: {e}"))?;
-
-    let mut measured = None;
-    while Instant::now() < deadline {
-        // If the app exited (e.g. the chosen port is already in use → bind
-        // failure → process exit), stop immediately so we never record a 200
-        // served by some other process listening on this port.
-        if let Ok(Some(status)) = child.try_wait() {
-            return Err(format!(
-                "scaffolded server exited before serving (exit: {status}); \
-                 port {port} may already be in use"
-            ));
-        }
-        // Accept only a 200 whose body proves it came from OUR scaffolded app:
-        // its index page renders the project name. This rules out a 200 served by
-        // some unrelated service that happens to answer on this port during the
-        // child's boot-toward-bind-failure window.
-        if let Ok(resp) = client.get(&url).send()
-            && resp.status().is_success()
-            && resp.text().is_ok_and(|body| body.contains(project_name))
-        {
-            measured = Some(u64::try_from(start.elapsed().as_millis()).unwrap_or(u64::MAX));
-            break;
-        }
-        std::thread::sleep(Duration::from_millis(100));
-    }
-
-    // Graceful shutdown so the app's on-shutdown hooks run — notably the
-    // managed-Postgres provider stopping its supervised cluster on the DB shape,
-    // which otherwise orphans Postgres processes/data between samples.
-    stop_child(&mut child);
-
-    measured.ok_or_else(|| "server did not return HTTP 200 before the deadline".to_string())
-}
-
-/// Reserve a currently-free TCP port on loopback.
-///
-/// Binds an ephemeral port, reads it, then drops the listener so the scaffolded
-/// app can bind it. There is a tiny race between release and the child's bind,
-/// but the per-iteration `try_wait()` and the response-body check make a
-/// mis-attributed sample impossible even if the port is lost.
-fn reserve_free_port() -> Result<u16, String> {
-    let listener = std::net::TcpListener::bind(("127.0.0.1", 0))
-        .map_err(|e| format!("failed to reserve a free port: {e}"))?;
-    let port = listener
-        .local_addr()
-        .map_err(|e| format!("failed to read reserved port: {e}"))?
-        .port();
-    Ok(port)
-}
-
 /// Extract the built binary path from `cargo build --message-format=json` output.
 ///
 /// Scans the `compiler-artifact` messages for the one whose target is the
 /// project's binary and returns its reported `executable` path. This is robust
 /// to a non-default target dir or `--target <triple>` configured via Cargo
 /// config files (which would otherwise move the artifact out of `target/debug/`).
-fn cargo_executable_path(stdout: &[u8], bin_name: &str) -> Option<PathBuf> {
+pub fn cargo_executable_path(stdout: &[u8], bin_name: &str) -> Option<PathBuf> {
     let text = String::from_utf8_lossy(stdout);
     let mut found = None;
     for line in text.lines() {
@@ -932,130 +699,6 @@ fn cargo_executable_path(stdout: &[u8], bin_name: &str) -> Option<PathBuf> {
         }
     }
     found
-}
-
-/// Stop a scaffolded cold-start server, preferring a graceful SIGTERM so the
-/// app runs its shutdown hooks (e.g. stopping a bundled Postgres cluster) before
-/// falling back to SIGKILL if it does not exit in time.
-fn stop_child(child: &mut std::process::Child) {
-    #[cfg(unix)]
-    {
-        if let Some(pid) = crate::process::validate_pid_for_kill(child.id())
-            && let Err(e) = nix::sys::signal::kill(
-                nix::unistd::Pid::from_raw(pid),
-                nix::sys::signal::Signal::SIGTERM,
-            )
-        {
-            eprintln!("  Warning: failed to SIGTERM cold-start server: {e}");
-        }
-        if crate::process::wait_with_timeout(child, Duration::from_secs(10)).is_err() {
-            let _ = child.kill();
-            let _ = child.wait();
-        }
-    }
-    #[cfg(not(unix))]
-    {
-        let _ = child.kill();
-        let _ = child.wait();
-    }
-}
-
-/// Run the cold-start measurement `runs` times, returning the samples (ms).
-fn measure_cold_start(shape: ColdStartShape, runs: u32) -> Result<Vec<u64>, String> {
-    let mut samples = Vec::with_capacity(runs as usize);
-    for i in 1..=runs {
-        eprintln!("  cold-start {shape:?} run {i}/{runs}…");
-        let ms = measure_cold_start_once(shape)?;
-        eprintln!("    → {ms} ms");
-        samples.push(ms);
-    }
-    Ok(samples)
-}
-
-/// Run the `autumn dev-loop-bench --cold-start` command.
-///
-/// In `--dry-run` mode it prints the cold-start budget table with no build or
-/// server. Otherwise it measures the no-DB `hello` cold start (gated) and,
-/// when `include_db` is set, the database-backed shape (informational).
-// Mirrors the flag set of [`run`] (json / fail_on_regression / dry_run) plus
-// the cold-start-only `include_db` toggle; grouping these into a struct would
-// not improve the single CLI call site.
-#[allow(clippy::fn_params_excessive_bools)]
-pub fn run_cold_start(
-    runs: u32,
-    output: Option<&str>,
-    json: bool,
-    fail_on_regression: bool,
-    dry_run: bool,
-    include_db: bool,
-) -> i32 {
-    if dry_run {
-        print!("{}", format_cold_start_budget_table(include_db));
-        return 0;
-    }
-
-    // A measurement run with zero samples would publish an all-zero, "passing"
-    // report without compiling or starting anything. Refuse it so the gate can
-    // never go green on no data.
-    if runs == 0 {
-        eprintln!("Error: --runs must be at least 1 for a cold-start measurement.");
-        return 1;
-    }
-
-    eprintln!(
-        "autumn dev-loop-bench --cold-start: measuring onboarding cold start ({runs} run(s))"
-    );
-    eprintln!(
-        "This scaffolds throwaway projects and compiles them from a cold target — \
-         expect minutes per run.\n"
-    );
-
-    let hello_samples = match measure_cold_start(ColdStartShape::Hello, runs) {
-        Ok(s) => s,
-        Err(e) => {
-            eprintln!("Error: cold-start (hello) measurement failed: {e}");
-            return 1;
-        }
-    };
-
-    let db = if include_db {
-        match measure_cold_start(ColdStartShape::Db, runs) {
-            Ok(s) => DbOutcome::Measured(s),
-            Err(e) => {
-                // The DB shape is informational; a failure must not break the run,
-                // but it is recorded in the report rather than silently dropped.
-                eprintln!("Warning: informational DB cold-start measurement failed: {e}");
-                DbOutcome::Failed(e)
-            }
-        }
-    } else {
-        DbOutcome::NotRequested
-    };
-
-    let report = build_cold_start_report(&hello_samples, &db);
-    let human = format_human_summary(&report);
-    let machine = format_json_report(&report);
-
-    if json {
-        println!("{machine}");
-    } else {
-        println!("{human}");
-    }
-
-    if let Some(path) = output {
-        if let Err(e) = std::fs::write(path, &machine) {
-            eprintln!("Warning: could not write report to {path}: {e}");
-        } else {
-            eprintln!("Report written to {path}");
-        }
-    }
-
-    if fail_on_regression && !report.all_passed {
-        eprintln!("Cold-start onboarding budget exceeded. Exiting 1.");
-        return 1;
-    }
-
-    0
 }
 
 // ── Tests ────────────────────────────────────────────────────────────────────
@@ -1818,40 +1461,49 @@ mod tests {
     }
 
     #[test]
-    fn reserve_free_port_returns_nonzero() {
-        let port = reserve_free_port().expect("should reserve a port");
-        assert!(port > 0);
+    fn cargo_executable_path_skips_non_json_lines() {
+        // Cargo may interleave non-JSON warning lines; they must be skipped, not
+        // abort parsing, so a later valid artifact line is still found.
+        let stdout = concat!(
+            "warning: some non-json diagnostic line\n",
+            r#"{"reason":"compiler-artifact","target":{"name":"coldstart_app"},"executable":"/tmp/x/target/debug/coldstart_app"}"#,
+            "\n",
+        );
+        let path = cargo_executable_path(stdout.as_bytes(), "coldstart_app");
+        assert_eq!(
+            path,
+            Some(PathBuf::from("/tmp/x/target/debug/coldstart_app"))
+        );
     }
 
-    // ── run_cold_start ────────────────────────────────────────────────────
+    // ── build_diagnostics next-action coverage ────────────────────────────
 
     #[test]
-    fn run_cold_start_dry_run_returns_zero() {
-        assert_eq!(run_cold_start(1, None, false, false, true, false), 0);
-        assert_eq!(run_cold_start(1, None, true, false, true, true), 0);
-    }
-
-    #[test]
-    fn run_cold_start_rejects_zero_runs() {
-        // --runs 0 must not publish an all-zero "passing" report without
-        // measuring anything; it returns a non-zero exit instead.
-        let exit = run_cold_start(0, None, false, false, false, false);
-        assert_eq!(exit, 1, "zero runs must be rejected");
-    }
-
-    #[test]
-    fn run_cold_start_dry_run_ignores_zero_runs() {
-        // Dry-run never measures, so runs == 0 is harmless there.
-        assert_eq!(run_cold_start(0, None, false, false, true, false), 0);
-    }
-
-    // ── live driver (slow: compiles a fresh project) ──────────────────────
-
-    #[test]
-    #[ignore = "compiles a throwaway project from a cold target; run with --ignored"]
-    fn cold_start_live_hello_measures_a_positive_duration() {
-        let ms = measure_cold_start_once(ColdStartShape::Hello)
-            .expect("live cold-start measurement should succeed");
-        assert!(ms > 0, "measured cold-start duration must be positive");
+    fn build_diagnostics_names_an_action_for_every_warm_class() {
+        // Every change class must yield a non-empty, actionable next step when it
+        // regresses — exercises each match arm in `build_diagnostics`.
+        for class in [
+            ChangeClass::InitialBoot,
+            ChangeClass::RustRouteEditHello,
+            ChangeClass::RustRouteEditDb,
+            ChangeClass::CssTailwind,
+            ChangeClass::StaticAsset,
+            ChangeClass::ConfigEdit,
+            ChangeClass::WatchDirEdit,
+        ] {
+            let budget = budget_for(class);
+            // Stats that blow past both the p95 and max budgets so the result fails
+            // and a diagnosis + next action are produced.
+            let stats = compute_stats(&[budget.max_ms * 4]);
+            let result = check_budget(class, stats, &budget);
+            assert!(
+                !result.passed,
+                "{class:?} should fail this over-budget input"
+            );
+            assert!(
+                !result.next_action.is_empty(),
+                "{class:?} must propose a next action"
+            );
+        }
     }
 }
