+++ codecov.yml
@@ -18,6 +18,7 @@ ignore:
   - "examples/**"
   - "autumn-cli/src/build.rs"   # CLI subprocess orchestration — untestable in unit tests
   - "autumn-cli/src/dev.rs"     # CLI subprocess orchestration — untestable in unit tests
+  - "autumn-cli/src/cold_start_driver.rs"  # Cold-start onboarding measurement driver: scaffold + cold cargo build + server spawn + HTTP poll — subprocess/TCP/filesystem orchestration, untestable in unit tests (like dev.rs/build.rs). Pure budget/report logic lives in dev_loop_bench.rs and is unit-tested there.
   - "autumn-cli/src/doctor.rs"  # Mix of pure-function unit tests and IO-bound orchestration; IO paths (process spawning, TCP, filesystem) excluded like build.rs/dev.rs
   - "autumn-cli/src/config.rs"  # psql subprocess orchestration (run_set/unset/list/history/check_psql) — untestable in unit tests; URL-resolution logic is covered in separate tests
   - "autumn-cli/src/flags.rs"  # psql subprocess orchestration (enable/disable/set-rollout/allow) — same pattern as config.rs
